refactor: change RAM data type from byte[][] to int[][] across multiple classes

+++ machine/src/main/java/com/fpetrola/oozx/DefaultRAMHolder.java
@@ -20,9 +20,9 @@ package com.fpetrola.oozx;
 
 public class DefaultRAMHolder implements RAMHolder {
   // RAM array: 65 pages of 16KB each (from SPECTRUM_RAM_PAGES)
-  private final byte[][] RAM = new byte[Memory.SPECTRUM_RAM_PAGES][0x4000];
+  private final int[][] RAM = new int[Memory.SPECTRUM_RAM_PAGES][0x4000];
 
-  public byte[][] getRAM() {
+  public int[][] getRAM() {
     return RAM;
   }
 }
